+++ lib/src/widgets/pdf_viewer_params.dart
@@ -28,6 +28,8 @@ class PdfViewerParams {
     this.onInteractionStart,
     this.onInteractionUpdate,
     this.onDocumentChanged,
+    this.calculateInitialPageNumber,
+    this.onViewerReady,
     this.onPageChanged,
     this.getPageRenderingScale,
     this.scrollByMouseWheel = 0.2,
@@ -123,7 +125,20 @@ class PdfViewerParams {
   final GestureScaleUpdateCallback? onInteractionUpdate;
 
   /// Function to notify that the document is loaded/changed.
+  ///
+  /// The function is called even if the document is null (it means the document is unloaded).
+  /// If you want to be notified when the viewer is ready to interact, use [onViewerReady] instead.
   final PdfViewerDocumentChangedCallback? onDocumentChanged;
+
+  /// Function called when the viewer is ready.
+  ///
+  /// Unlike [PdfViewerDocumentChangedCallback], this function is called after the viewer is ready to interact.
+  final PdfViewerReadyCallback? onViewerReady;
+
+  /// Function to calculate the initial page number.
+  ///
+  /// It is useful when you want to determine the initial page number based on the document content.
+  final PdfViewerCalculateInitialPageNumberFunction? calculateInitialPageNumber;
 
   /// Function called when the current page is changed.
   final PdfPageChangedCallback? onPageChanged;
@@ -298,6 +313,8 @@ class PdfViewerParams {
         other.onInteractionStart == onInteractionStart &&
         other.onInteractionUpdate == onInteractionUpdate &&
         other.onDocumentChanged == onDocumentChanged &&
+        other.calculateInitialPageNumber == calculateInitialPageNumber &&
+        other.onViewerReady == onViewerReady &&
         other.onPageChanged == onPageChanged &&
         other.getPageRenderingScale == getPageRenderingScale &&
         other.scrollByMouseWheel == scrollByMouseWheel &&
@@ -331,6 +348,8 @@ class PdfViewerParams {
         onInteractionStart.hashCode ^
         onInteractionUpdate.hashCode ^
         onDocumentChanged.hashCode ^
+        calculateInitialPageNumber.hashCode ^
+        onViewerReady.hashCode ^
         onPageChanged.hashCode ^
         getPageRenderingScale.hashCode ^
         scrollByMouseWheel.hashCode ^
@@ -350,6 +369,21 @@ class PdfViewerParams {
 
 /// Function to notify that the document is loaded/changed.
 typedef PdfViewerDocumentChangedCallback = void Function(PdfDocument? document);
+
+/// Function to calculate the initial page number.
+///
+/// If the function returns null, the viewer will show the page of [PdfViewer.initialPageNumber].
+typedef PdfViewerCalculateInitialPageNumberFunction = int? Function(
+  PdfDocument document,
+  PdfViewerController controller,
+);
+
+/// Function called when the viewer is ready.
+///
+typedef PdfViewerReadyCallback = void Function(
+  PdfDocument document,
+  PdfViewerController controller,
+);
 
 /// Function called when the current page is changed.
 typedef PdfPageChangedCallback = void Function(int? pageNumber);